Change the type of fresh to put the Env and Fresh constraints on separate levels. The new type enforces a layering of Env on the top level and Fresh in the bottom.
This is what makes sense usually, as we want the global
computation to run inside the same Fresh monad, while the environment
is local to a fresh operation and should be closed afterwards.

I'm not entirely satisfied with the new type, as it looks more ad-hoc
and scary, but the old type was making things ridiculously hard for the
client to use fresh.